pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "live-ml-project"
version = "0.1.0"
description = "Live financial market prediction system with online learning"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "joblib>=1.3.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "river>=0.21.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "mlflow>=2.16.0",
    "great-expectations>=0.18.0",
    "deepchecks>=0.18.0",
    "google-cloud-logging>=3.9.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-secret-manager>=2.18.0",
    "google-cloud-aiplatform>=1.43.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "dvc>=3.42.0",
    "dvc-gs>=3.0.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"
